Exrtract VisitBadge into new sub component
