+++ Examples/Widgets/CRM/src/components/VisitBadge.scss
@@ -1,0 +1,158 @@
+/**
+ * Copyright (c) Cisco Systems, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+@import "./TableStyles";
+
+:host {
+  width: 100vw;
+}
+
+.time {
+  display: flex;
+  flex-direction: column-reverse;
+  justify-content: space-around;
+  grid-area: time;
+}
+
+.months {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  grid-area: month;
+}
+
+.year {
+  grid-area: year;
+}
+
+.months-line {
+  align-content: center;
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  padding-right: 1rem;
+  margin: 1rem 0;
+}
+
+div.month-line {
+  background: #dedede;
+  content: " ";
+  display: flex;
+  flex-grow: 1;
+  height: 0.125rem;
+  margin: 0;
+  position: relative;
+}
+
+div.month-line::before {
+  background: white;
+  border: 2px solid #dedede;
+  content: " ";
+  border-radius: 50%;
+  height: 0.625rem;
+  width: 0.625rem;
+  align-self: center;
+  position: absolute;
+}
+
+div.month-line:last-child {
+  flex: 0 1 0;
+}
+
+.month-names {
+  display: flex;
+  justify-content: space-between;
+}
+
+md-badge {
+  box-shadow: 0px 0px 1px 0px #000000 18%;
+  box-shadow: 0px 2px 4px 0px #000000 16%;
+}
+
+.mock-badge {
+  display: flex;
+  height: 1rem;
+  justify-content: flex-start;
+  align-items: center;
+  border-radius: 1.5rem;
+  margin-bottom: auto;
+  overflow: hidden;
+  text-align: center;
+  text-decoration: none;
+  white-space: nowrap;
+  z-index: 100;
+  font-size: 0.625rem;
+  line-height: normal;
+  padding: 0.25rem 1rem 0.25rem 0;
+  background-color: #ffffff;
+  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.16), 0px 0px 1px rgba(0, 0, 0, 0.18);
+  cursor: pointer;
+
+  &:hover {
+    background-color: #e6fbff;
+  }
+
+  .split-left {
+    display: flex;
+    align-items: center;
+  }
+  .split-separator {
+    margin: 0 0.25rem;
+  }
+}
+
+md-menu-overlay::part(overlay-content) {
+  width: 16.5rem;
+}
+
+.visit-overlay-content {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.visit-details-header {
+  border-bottom: 1px solid #cccccc;
+  padding: 1rem;
+  display: flex;
+  justify-content: space-around;
+  align-items: flex-start;
+
+  md-badge {
+  }
+
+  &:nth-child(2) {
+    width: 80%;
+    text-align: start;
+  }
+
+  h2, h3  {
+    margin: 0;
+  }
+
+  h2 {
+    font-size: .875rem;
+  }
+  
+  h3 {
+    font-size: .625rem;
+    font-weight: lighter;
+    color: var(--primary-text-color, #777);
+  }
+
+}
+
+.visit-details-content {
+  // width: 100%;
+  padding: 1rem;
+}
+
+table {
+  @include table;
+}
